requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
openpyxl>=3.1.0
python-docx>=1.1.0
calamine>=0.1.7

